requirements.txt
alembic==0.8.10
appdirs==1.4.0
blinker==1.3
click==6.7
Faker==0.7.7
Flask==0.12
Flask-Assets==0.12
Flask-Compress==1.4.0
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-RQ==0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-SSLify==0.1.5
Flask-WTF==0.11
gunicorn==19.6.0
honcho==0.7.1
itsdangerous==0.24
Jinja2==2.9.5
jsmin==2.2.1
jsonpickle==0.9.2
Mako==1.0.6
MarkupSafe==0.23
packaging==16.8
psycopg2==2.7
pyparsing==2.1.10
python-dateutil==2.6.0
python-editor==1.0.3
raygun4py==3.1.4
redis==2.10.5
requests==2.9.1
rq==0.7.1
six==1.10.0
SQLAlchemy==1.1.5
webassets==0.12.1
Werkzeug==0.11.15
WTForms==2.1
#git+https://github.com/mobolic/facebook-sdk.git#egg=facebook-sdk
pytz
